Add listener meta-inf file
This file when placed in the resource folder will register the listener.
It cannot be placed in the resource folder of this project because
then it will activate when the tests are ran. Because of that it is
placed in a separate folder where it will remain inactive for the
current project but it will be copied in the jar artifact when building
the project.